requirements.txt
nginx
starlette
uvicorn
llama-cpp-python
